[FE] rename: button 폴더 이름 변경
why: button 폴더 이름을 변경했다. 그냥 button이라는 폴더 이름은 혼란을
야기할 수 있기에, 프로젝트에 맞춘 button이라는 사실을 강조하기 위하여,
customButton아ㅣ라고 이름을 교체하였다.
